# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": " Questioner API helps meetup organizer prioritize questions and answers. ",
    "version": "1.0.0",
    "title": "Questioner",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "info@questioner.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "seyiproject.herokuapp.com",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "Meetups",
      "description": "Post, Get and Delete Meetup"
    },
    {
      "name": "Questions",
      "description": "Post and Get Question"
    },
    {
      "name": "Comments",
      "description": "Post and Get Comment"
    },
    {
      "name": "Users",
      "description": "SignUp and Login"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/meetups": {
      "post": {
        "tags": [
          "Meetups"
        ],
        "summary": "Add a new meetup",
        "description": "This will add new meetup resource",
        "operationId": "createMeetup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Meetup Details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createMeetup"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Meetup Successfully Created",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication Errors"
          },
          "403": {
            "description": "Unauthorized User"
          }
        },
        "security": [
          {
            "tokens": [

            ]
          }
        ]
      },
      "get": {
        "tags": [
          "Meetups"
        ],
        "summary": "Get All Meetups",
        "description": "",
        "operationId": "getAllMeetupsRecord",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation Or No Meetups Exist",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/meetups/upcoming": {
      "get": {
        "tags": [
          "Meetups"
        ],
        "summary": "Get upcoming meeetup",
        "description": "",
        "operationId": "upcomingMeetups",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": [

            ]
          }
        ]
      }
    },
    "/meetups/{meetupId}": {
      "get": {
        "tags": [
          "Meetups"
        ],
        "summary": "Get a specific Meetup by Id",
        "description": "",
        "operationId": "getASpecificMeetupRecord",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "meetupId",
            "in": "path",
            "description": "Gets meetup by id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operations"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": [

            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Meetups"
        ],
        "summary": "Delete a Meetup",
        "description": "",
        "operationId": "deleteMeetup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "meetupId",
            "in": "path",
            "description": "meetup id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "meetup successfuly deleted"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": [

            ]
          }
        ]
      }
    },
    "/meetups/{meetupId}/rsvps": {
      "post": {
        "tags": [
          "Meetups"
        ],
        "summary": "Add new Rsvp to Meetup",
        "description": "",
        "operationId": "createRsvps",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tokens",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "meetupId",
            "in": "path",
            "description": "Gets meetup by id",
            "type": "integer",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Gets meetup by id",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RSVP"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": [

            ]
          }
        ]
      }
    },
    "/questions": {
      "post": {
        "tags": [
          "Questions"
        ],
        "summary": "Add new question",
        "description": "",
        "operationId": "createQuestion",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tokens",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createQuestion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "question successfully added"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": [

            ]
          }
        ]
      }
    },
    "/meetups/:meetupId/questions": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get questions of a meetup",
        "description": "",
        "operationId": "getASpecificQuestionRecord",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tokens",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "meetupId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/questions/{questionId}/upvote": {
      "patch": {
        "tags": [
          "Questions"
        ],
        "summary": "Upvote a question",
        "description": "",
        "operationId": "upvote",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tokens",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation Or Meetup-Id Doesnt Exit"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": [

            ]
          }
        ]
      }
    },
    "/questions/{questionId}/downvote": {
      "patch": {
        "tags": [
          "Questions"
        ],
        "summary": "Down vote  a question",
        "description": "",
        "operationId": "downvote",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tokens",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation Or Meetup-Id Doesnt Exit"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": [

            ]
          }
        ]
      }
    },
    "/comments": {
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Posting of Comment on a Question",
        "description": "",
        "operationId": "addComment",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tokens",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication Errors"
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "tokens": [

            ]
          }
        ]
      }
    },
    "/questions/:questionId/comments": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Gets a Comments by Question Id",
        "description": "",
        "operationId": "getComments",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "403": {
            "description": "forbidden token access"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user",
        "description": "",
        "operationId": "userSignUp",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "403": {
            "description": "Token Forbidden"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "User Login",
        "description": "",
        "operationId": "userLogin",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "User Name",
            "required": true,
            "type": "string",
            "format": "email"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "User Name",
            "required": true,
            "type": "string",
            "format": "password"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "401": {
            "description": "Invalid Login credential"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "tokens": {
      "type": "apiKey",
      "name": "token",
      "in": "header"
    }
  },
  "definitions": {
    "createMeetup": {
      "type": "object",
      "required": [
        "topic",
        "location",
        "happeningOn"
      ],
      "properties": {
        "topic": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "happeningOn": {
          "type": "string",
          "format": "data-time"
        }
      },
      "xml": {
        "name": "Meetup"
      }
    },
    "createQuestion": {
      "type": "object",
      "required": [
        "meetup",
        "title"
      ],
      "properties": {
        "meetup": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Question"
      }
    },
    "RSVP": {
      "type": "object",
      "required": [
        "response"
      ],
      "properties": {
        "response": {
          "type": "string"
        }
      },
      "xml": {
        "name": "RSVP"
      }
    },
    "createUser": {
      "type": "object",
      "required": [
        "firstname",
        "lastname",
        "othername",
        "email",
        "phoneNumber",
        "userName",
        "password"
      ],
      "properties": {
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "format": "password"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer"
        },
        "data": {
          "type": "array",
          "xml": {
            "name": "tag",
            "wrapped": true
          },
          "items": {
            "type": "string"
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}